+++ src/main.js
@@ -6,6 +6,8 @@ import '@/style/common.less'
 import '@/utils/vant-ui'
 
 import Modal from '@/utils/modal'
+import loadingDirective from './utils/loading'
+Vue.directive('loading', loadingDirective)
 Vue.use(Modal)
 
 Vue.config.productionTip = false