+++ src/utils/vant-ui.js
@@ -1,4 +1,5 @@
 import Vue from 'vue'
-import { Rate, Icon } from 'vant'
+import { Toast, Rate, Icon } from 'vant'
+Vue.use(Toast)
 Vue.use(Rate)
 Vue.use(Icon)